+++ babylon/scaffold/src/app/babylon.ts
@@ -1,0 +1,36 @@
+export { Engine } from "@babylonjs/core/Engines/engine";
+export { Scene } from "@babylonjs/core/scene";
+
+export { Vector2, Vector3, Quaternion, Matrix } from "@babylonjs/core/Maths/math.vector";
+export { Color3, Color4 } from "@babylonjs/core/Maths/math.color";
+
+export { ArcRotateCamera } from "@babylonjs/core/Cameras/arcRotateCamera";
+export { FreeCamera } from "@babylonjs/core/Cameras/freeCamera";
+export { UniversalCamera } from "@babylonjs/core/Cameras/universalCamera";
+
+export { HemisphericLight } from "@babylonjs/core/Lights/hemisphericLight";
+export { DirectionalLight } from "@babylonjs/core/Lights/directionalLight";
+export { PointLight } from "@babylonjs/core/Lights/pointLight";
+export { SpotLight } from "@babylonjs/core/Lights/spotLight";
+
+export { StandardMaterial } from "@babylonjs/core/Materials/standardMaterial";
+export { PBRMaterial } from "@babylonjs/core/Materials/PBR/pbrMaterial";
+export { Texture } from "@babylonjs/core/Materials/Textures/texture";
+export { CubeTexture } from "@babylonjs/core/Materials/Textures/cubeTexture";
+
+export { Mesh } from "@babylonjs/core/Meshes/mesh";
+export { AbstractMesh } from "@babylonjs/core/Meshes/abstractMesh";
+
+export { CreateBox } from "@babylonjs/core/Meshes/Builders/boxBuilder";
+export { CreateSphere } from "@babylonjs/core/Meshes/Builders/sphereBuilder";
+export { CreateGround } from "@babylonjs/core/Meshes/Builders/groundBuilder";
+export { CreatePlane } from "@babylonjs/core/Meshes/Builders/planeBuilder";
+export { CreateCylinder } from "@babylonjs/core/Meshes/Builders/cylinderBuilder";
+export { CreateTorus } from "@babylonjs/core/Meshes/Builders/torusBuilder";
+
+export { SceneLoader } from "@babylonjs/core/Loading/sceneLoader";
+export { Animation } from "@babylonjs/core/Animations/animation";
+export { Sound } from "@babylonjs/core/Audio/sound";
+
+export { ActionManager } from "@babylonjs/core/Actions/actionManager";
+export { ExecuteCodeAction } from "@babylonjs/core/Actions/directActions";